moved every component into their own folder. Stateful components will be part of folder hierarchy, stateless components will hold other folders.
