+++ src/App.css
@@ -3,6 +3,7 @@
 	--game-max-width: 500px;
 	--keyboard-height: 200px;
 	--key-color: #818384;
+	--color-tone-1: #ffffff;
 }
 html {
 	background-color: #121213;
