+++ src/App.js
@@ -1,7 +1,7 @@
 import './App.css';
 
-import Header from './components/header/Header';
-import Board from './components/board/Board'
+import Header from './components/Header/Header';
+import Board from './components/Board/Board'
 
 
 function App() {
