+++ src/components/board/Tile/Tile.css
@@ -10,6 +10,7 @@
 	box-sizing: border-box;
 	text-transform: uppercase;
 	user-select: none;
+	cursor: pointer;
 }
 .highlighted {
 	border: 2px solid white !important;
