+++ src/components/board/keyboard/Keyboard.css
@@ -2,3 +2,6 @@
   margin: 0 8px;
   user-select: none;
 }
+.keyboard-container {
+	height: var(--keyboard-height);
+}
