+++ src/components/header/Help/Help.css
@@ -1,0 +1,4 @@
+.HelpSvg {
+  margin-left: 10px;
+  cursor: pointer;
+}
